chore : entity 구조 변경
Id값을 모두 Long타입으로 변경했습니다.

+++ src/main/java/com/book_everywhere/domain/pin/PinRepository.java
@@ -2,5 +2,5 @@ package com.book_everywhere.domain.pin;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PinRepository extends JpaRepository<Pin, Integer> {
+public interface PinRepository extends JpaRepository<Pin, Long> {
 }
